Added Native Android App
Added eclipse project that contains the native android app. Currently
this contains the map and videos all with gui. It links to the existing
wikitude app which contains the pavilion models.

It is intended to eventually contain the audio files and the logic that
controls whether users are "at" a pavilion (and thus can see the
models, video, and audio).