Location Awareness for Map/GUI
Added location awareness and logic. App now only displays gui  (and
thus access to additional components) when user is within close range
of the specified locations at the World’s Fair site. It plays the video
appropriate for each site. (It does not play a video for the African
pavilion as that is not available yet.) Location is updated roughly
every five seconds for purposes of determining if user has entered or
exited the range of one of the locations.